+++ src/main/java/net/minecraftforge/common/ForgeConfig.java
@@ -222,91 +222,4 @@ public class ForgeConfig {
     //public static boolean logCascadingWorldGeneration = true; // see Chunk#logCascadingWorldGeneration()
     //public static boolean fixVanillaCascading = false;
 
-    public static class Common {
-
-        Common(ForgeConfigSpec.Builder builder) {
-            builder.comment("a Boolean comment")
-                    .translation("forge.configgui.a.boolean")
-                    .worldRestart()
-                    .define("aBoolean", false);
-
-            builder.comment("Category comment")
-                    .push("numbers");
-            {
-                builder.comment("an Int comment")
-                        .translation("forge.configgui.an.int")
-                        .defineInRange("anInt", 5, -10, 1000);
-                builder.comment("a Long comment")
-                        .translation("forge.configgui.a.long")
-                        .defineInRange("aLong", Long.MIN_VALUE, Long.MIN_VALUE, 5);
-                builder.comment("a Double comment")
-                        .translation("forge.configgui.a.double")
-                        .defineInRange("aDouble", 5d, 4d, 6d);
-
-                builder.comment("Sub-category comment")
-                        .push("empty");
-                builder.pop();
-
-                builder.comment("Sub-category comment")
-                        .push("not-empty");
-                {
-                    builder.comment("sub - an Int comment")
-                            .translation("forge.configgui.sub.an.int")
-                            .defineInRange("subAnInt", 5, -10, 1000);
-                }
-                builder.pop();
-            }
-            builder.pop();
-
-            builder.comment("an Enum comment")
-                    .translation("forge.configgui.an.enum")
-                    .defineEnum("anEnum", DyeColor.GREEN, dc -> dc instanceof DyeColor && ((DyeColor) dc).getId() >= 10);
-            builder.comment("a String comment")
-                    .translation("forge.configgui.a.string")
-                    .define("aString", "foo");
-            builder.comment("an Enum List comment")
-                    .translation("forge.configgui.an.enum.list")
-                    .defineList("anEnumList", Lists.newArrayList(), o -> o instanceof String);
-            builder.comment("a String In List comment")
-                    .translation("forge.configgui.a.string.in.list")
-                    .defineInList("aStringInList", "bar", Lists.newArrayList("foo", "bar", "baz"));
-
-            builder.comment("I'm in Spain but the S is silent")
-                    .push("Lists!");
-            {
-                builder.comment("a List<Boolean> comment")
-                        .translation("forge.configgui.a.list.boolean")
-                        .defineList("aListBoolean", Lists.newArrayList(true, false, true, false, true, true, false, false), o -> o instanceof Boolean);
-                builder.comment("a List<Int> comment")
-                        .translation("forge.configgui.a.list.int")
-                        .defineList("aListInt", Lists.newArrayList(1, 2, 5, 66, 3, 7), o -> o instanceof Integer);
-                builder.comment("a List<Long> comment")
-                        .translation("forge.configgui.a.list.long")
-                        .defineList("aListLong", Lists.newArrayList(5L, 12L, 623L, Long.MIN_VALUE), o -> o instanceof Long);
-                builder.comment("a List<Double> comment")
-                        .translation("forge.configgui.a.list.double")
-                        .defineList("aListDouble", Lists.newArrayList(123.213, 21343.21, 456.32, 9765.2), o -> o instanceof Double);
-                builder.comment("a List<Enum> comment")
-                        .translation("forge.configgui.a.list.enum")
-                        .defineList("aListEnum", Lists.newArrayList(DyeColor.BLACK, DyeColor.GREEN, DyeColor.RED, DyeColor.BLACK), o -> o instanceof DyeColor);
-                builder.comment("a List<String> comment")
-                        .translation("forge.configgui.a.list.string")
-                        .defineList("aListString", Lists.newArrayList("foo", "bar", "baz", "qez"), o -> o instanceof String);
-                builder.comment("a List<List<String>> comment")
-                        .translation("forge.configgui.a.list.list.string")
-                        .defineList("aListListString", Lists.newArrayList(Lists.newArrayList("foo", "bar", "baz", "qez"), Lists.newArrayList("foobar")), o -> o instanceof List);
-            }
-            builder.pop();
-        }
-    }
-
-    static final ForgeConfigSpec commonSpec;
-    public static final Common COMMON;
-
-    static {
-        final Pair<Common, ForgeConfigSpec> specPair = new ForgeConfigSpec.Builder().configure(Common::new);
-        commonSpec = specPair.getRight();
-        COMMON = specPair.getLeft();
-    }
-
 }
